src/server/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-decouple==3.8
setuptools==75.5.0
sqlparse==0.5.1
tzdata==2024.2
